Move sequelize config to a separate service so it can be reused easily
