+++ libs/nest/sequelize/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "extends": ["../../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"],
+  "rules": {}
+}
