+++ libs/nest/sequelize/README.md
@@ -1,0 +1,7 @@
+# Nest Sequelize
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test nest-sequelize` to execute the unit tests via [Jest](https://jestjs.io).
