+++ libs/nest/sequelize/jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  displayName: 'nest-sequelize',
+  preset: '../../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsConfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../../coverage/libs/nest/sequelize',
+}
